+++ service/src/model/page.rs
@@ -4,7 +4,7 @@ pub struct Page<T> {
     /// The actual entries on the page
     pub entries: Vec<T>,
     /// The offset of this page in the total resultset
-    pub offset: u32,
+    pub offset: u64,
     /// The total number of records in the resultset
-    pub total: u32,
+    pub total: u64,
 }
